setuassignment/requirements.txt
Django==1.9.2
MySQL-python==1.2.5
PyYAML==3.12
django-extensions==1.6.1
django-rest-swagger==0.3.5
requests==2.11.1
djangorestframework==3.3.2
djangorestframework-api-key